いいね関連のメソッドをmixin化
