+++ front/pages/index.vue
@@ -25,11 +25,13 @@ import { mapGetters } from 'vuex';
 import SelectCategory from '@/components/Category/SelectCategory.vue';
 import { handleFailure } from '@/plugins/error-handler';
 import SampleList from '@/components/Sample/SampleList.vue';
+import likeMixin from '@/mixins/likeMixin.js';
 export default {
   components: {
     SelectCategory,
     SampleList,
   },
+  mixins: [likeMixin],
   data() {
     return {
       isLoading: false,
@@ -86,48 +88,6 @@ export default {
       } catch (error) {
         handleFailure(error, this.$store);
       }
-    },
-    async addLike(sampleId) {
-      if (this.isLoading) return;
-      this.isLoading = true;
-      try {
-        await this.$axios.$post('/api/v1/likes/', {
-          like: {
-            sample_id: sampleId,
-            user_id: this.$auth.user.id,
-          },
-        });
-        await this.getSamples();
-      } catch (error) {
-        handleFailure(error, this.$store);
-      } finally {
-        this.isLoading = false;
-      }
-    },
-    findLikeId(sample) {
-      const likeObject = sample.likes.find(
-        (like) => like.user_id === this.$auth.user.id,
-      );
-      return likeObject ? likeObject.id : null;
-    },
-    async deleteLike(likeId) {
-      if (this.isLoading) return;
-      this.isLoading = true;
-      try {
-        await this.$axios.$delete(`/api/v1/likes/${likeId}`);
-        await this.getSamples();
-      } catch (error) {
-        handleFailure(error, this.$store);
-      } finally {
-        this.isLoading = false;
-      }
-    },
-    notLoginUserClick() {
-      this.$store.dispatch('getToast', {
-        msg: 'ログインユーザのみいいねをつけることができます',
-        color: 'primary',
-        timeout: 4000,
-      });
     },
   },
 };
